moved tx toggle trad controller
